memory_agents/customer_support_voice_agent/requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.25
memori>=0.3.0
openai>=1.40.0
firecrawl-py>=0.1.0
pydantic>=2.7.0


